refactor: invite 폴더 도메인과 컴포넌트 분리
- full-layout으로 이동
- 기존 app/ 폴더 안에 있던 invite 하위 폴더를 app 바깥으로 이동
- app/invite/ 폴더는 page.tsx만 존재
- import 경로 수정
